+++ GUI_interpilacion_newton_3.py
@@ -71,7 +71,7 @@ class Gui_interpolacionNewton3:
                 if x1 == "":
                     messagebox.showerror("Interpolacion de Newton","Error: El primer punto en su campo x se encuentra vacío")
                 else:
-                    x1 = int(x1)
+                    x1 = float(x1)
                     c1 = True
             except ValueError:
                 messagebox.showerror("Interpolacion de Newton","Error: Debe ingresar solo valores numéricos.")
@@ -81,7 +81,7 @@ class Gui_interpolacionNewton3:
                 if x2 == "":
                     messagebox.showerror("Interpolacion de Newton","Error: El primer punto en su campo f(x) se encuentra vacío")
                 else:
-                    x2 = int(x2)
+                    x2 = float(x2)
                     c2 = True
             except ValueError:
                 messagebox.showerror("Interpolacion de Newton","Error: Debe ingresar solo valores numéricos.")
@@ -91,7 +91,7 @@ class Gui_interpolacionNewton3:
                 if y1 == "":
                     messagebox.showerror("Interpolacion de Newton","Error: El segundo punto en su campo x se encuentra vacío")
                 else:
-                    y1 = int(y1)
+                    y1 = float(y1)
                     c3 = True
             except ValueError:
                 messagebox.showerror("Interpolacion de Newton","Error: Debe ingresar solo valores numéricos.")
@@ -101,7 +101,7 @@ class Gui_interpolacionNewton3:
                 if y2 == "":
                     messagebox.showerror("Interpolacion de Newton","Error: El segundo punto en su campo f(x) se encuentra vacío")
                 else:
-                    y2 = int(y2)
+                    y2 = float(y2)
                     c4 = True
             except ValueError:
                 messagebox.showerror("Interpolacion de Newton","Error: Debe ingresar solo valores numéricos.")
@@ -111,7 +111,7 @@ class Gui_interpolacionNewton3:
                 if z1 == "":
                     messagebox.showerror("Interpolacion de Newton","Error: El tercer punto en su campo x se encuentra vacío")
                 else:
-                    z1 = int(z1)
+                    z1 = float(z1)
                     c5 = True
             except ValueError:
                 messagebox.showerror("Interpolacion de Newton","Error: Debe ingresar solo valores numéricos.")
@@ -121,7 +121,7 @@ class Gui_interpolacionNewton3:
                 if z2 == "":
                     messagebox.showerror("Interpolacion de Newton","Error: El tercer punto en su campo f(x) se encuentra vacío")
                 else:
-                    z2 = int(z2)
+                    z2 = float(z2)
                     c6 = True
             except ValueError:
                 messagebox.showerror("Interpolacion de Newton","Error: Debe ingresar solo valores numéricos.")
@@ -130,7 +130,7 @@ class Gui_interpolacionNewton3:
                 if w1 == "":
                     messagebox.showerror("Interpolacion de Newton","Error: El Cuarto punto en su campo x se encuentra vacío")
                 else:
-                    w1 = int(w1)
+                    w1 = float(w1)
                     c7 = True
             except ValueError:
                 messagebox.showerror("Interpolacion de Newton","Error: Debe ingresar solo valores numéricos.")
@@ -140,7 +140,7 @@ class Gui_interpolacionNewton3:
                 if w2 == "":
                     messagebox.showerror("Interpolacion de Newton","Error: El Cuarto punto en su campo f(x) se encuentra vacío")
                 else:
-                    w2 = int(w2)
+                    w2 = float(w2)
                     c8 = True
             except ValueError:
                 messagebox.showerror("Interpolacion de Newton","Error: Debe ingresar solo valores numéricos.")
